requirements.txt
Django==1.11.18

django-filter==1.1
django-cors-headers==1.2.0



# REST
djangorestframework==3.8.2
djangorestframework-jwt==1.9.0
djoser==1.1.5

# dev
ipdb

uWSGI==2.0.15
dj-database-url==0.4.1
psycopg2==2.6.1

# testing
pytest-django==3.4.3
pytest-factoryboy==2.0.2
factory-boy>=2.10.0
Faker==1.0.0
pytest-cov

django-import-export==1.2.0
